+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 -----
 * Add ``add_users`` and ``remove_users`` method to Group resource
 * Add ``is_active`` field to ``Process`` resource
+* Add ``parents`` and ``children`` property to ``Data``
 
 Fixed
 -----